+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [1.0.0] - 2020-12-12.
+## 1.0.1(2020-12-14)
+
+* Add `showCountryAndCode` and `showCountryOnly` config
+* Add `listPageAppBarTheme` and `listPageAppBarTheme` customize config
+* Fix error in list page
+
+## 1.0.0(2020-12-12)
 
 * first release